Modernize codebase for C++17/20, improve type safety
Refactored core engine code for consistency and correctness:
- Replaced legacy casts with static_cast and updated template parameters to use SIZE_T for container/vector sizes.
- Added custom bit_cast for pre-C++20 compilers; updated float and int conversions for portability.
- Improved math utilities (repeat, ceilLog2, gridSnap) and rewrote quartic solver for accuracy.
- Updated hash functors to use local hashVal variables for clarity.
- Cleaned up Random class and removed scripting macros.
- Standardized memory allocation and placement new usage.
- Enhanced string formatting and data stream encoding handling.
- Rewrote Vector3 orientation conversion methods for correctness.
- Updated includes, macros, and test cases for modern C++ compatibility.
- Improved RTTR registration and test metadata handling.
These changes fix subtle bugs, improve maintainability, and ensure compatibility with newer C++ standards.

+++ sdk/geUtilities/include/geDebug.h
@@ -167,12 +167,12 @@ bool LogCategory##name::s_registered = Log::_registerCategory(LogCategory##name:
 #define GE_LOG(verbosity, category, message, ...)                             \
   do {                                                                        \
   using namespace ::geEngineSDK;                                              \
-  IF_CONSTEXPR ((int32)LogVerbosity::verbosity <= (int32)GE_LOG_VERBOSITY) {  \
+  IF_CONSTEXPR (int32(LogVerbosity::verbosity) <= int32(GE_LOG_VERBOSITY)) {  \
     g_debug().log(StringUtil::format(message, ##__VA_ARGS__) +                \
                           String("\n\t\t in ") +                              \
                           __PRETTY_FUNCTION__ +                               \
                           " [" + __FILE__ + ":" +                             \
-                          toString(static_cast<int32>(__LINE__)) + "]\n",     \
+                          toString(__LINE__) + "]\n",                         \
                   LogVerbosity::verbosity, LogCategory##category::_id);       \
   }} while (0)
